+++ src/wai/common/geometry/_Point.py
@@ -1,0 +1,7 @@
+class Point:
+    """
+    Represents an (x, y) coordinate in 2D integer Cartesian space.
+    """
+    def __init__(self, x: int, y: int):
+        self.x: int = x
+        self.y: int = y