+++ src/wai/common/geometry/__init__.py
@@ -1,0 +1,6 @@
+"""
+Package for working with shapes.
+"""
+from ._Point import Point
+from ._Polygon import Polygon
+from ._Rectangle import Rectangle